Revert "Remove unnecessary page-header class and define style for header element instead"
This reverts commit 6cb8cf97a28888774d3151e577d53fd91a9ff543.

Problem is header is used in My Projects, About me and other headings
too. We do not want to use same style for all headers everywhere as it will create
visual confusion.

My point is main header on page that shows page title should not be used
any where else on the same page so that it remains visually distinct.