Category under article name made optional
On the home page in the recent posts section the article category
was listed under the article name, for example:

	Installing python packages
	posted in misc									Jan 05 2015

If users are not using categories on their pelican website, than this
information is irrelevant. Now, the information is not displayed when
the variable `DISPLAY_CATEGORIES_ON_MENU` is unset.