ci(theme): use theme from the repo instead of downloading a fresh copy
When CI downloads a fresh copy to build, it is building against the
release version. This way even in next branch, we never get to see next
release in action.

Secondly, why download a fresh copy when you have already got it in the
repo.

This will save us in build time