fix(perf): remove a nested loop in the tag list generation
This patch removes the nested loop in the generation of the tag list
alongside an article. It makes uses of the fact that `tags` is a
dictionary.

On my computer this reduces the generation time of the documentation
site from 73s to 70s.

(This PR is related to #145. Even though it removes the nested loop,
I feel that it does not fully close the issue because the generation
performance remains low. Someone will need to profile this again and
find out what is the next bottleneck.)

Informs #145