feat(categories): add support for more_categories plugin
The `more_categories` plugin enables having more than
one category for a single article by separating them
with commas.
This commit comprises the required template changes
to display all related categories instead of just one.
The change is safe when the plugin is not used.
It changes the category list in the last article section
of the index and in the article page.
It also modifies the keyword meta, to use all categories.
This commit also fixes an unreported bug with the keyword meta:
when no keywords or tags are specified, pointless commas were
added at the beginning or the end.