Few minor corrections
Spell as "GoatCounter" consistently, and also rename the variable as a
single word as that's how it's spelled everywhere else (e.g. if people
want to add events or use window.goatcounter).

Also rewrite the text a bit: I added a text field to the top of the docs
for easier copy/pasting.

Include the protocol in "GOATCOUNTER" in case people want to self-host
over http (not recommended, but it works).