refactor(categories, comments): remove Bootstrap.js
Removing Bootstrap.js in the base.html requires a refactor to make the accordions on the categories
page and in the comment section work correctly. This commit refactores the accordion with a hidden
checkbox and some css. Aditionally some js was changed to allow anchorlinking to a specific category
and opening it.